Trim input value before adding tasks to prevent leading/trailing spaces
previously the leading or trailing whitespaces unintentionally added by the user would get saved in the local storage along with the input. By adding trim functionality the leading and trailing white spaces are taken care of. Also if search or filter functionality is implmented in the future without trim , it may cause inaccurate input matching.